requirements/base.txt
asgiref==3.7.2
astroid==2.15.6
async-timeout==4.0.3
dill==0.3.7
Django==4.2.5
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-model-utils==4.3.1
djangorestframework==3.14.0
hiredis==2.2.3
importlib-metadata==6.8.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==10.0.0
platformdirs==3.10.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pylint==2.17.5
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
singlemodeladmin==0.9
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
ua-parser==0.18.0
user-agents==2.2.0
wrapt==1.15.0
yapf==0.40.1
zipp==3.16.2
